Added "apply" command
close #22
