+++ README.rst
@@ -30,3 +30,10 @@ email and password from github.
 
     $ curl -X POST http://127.0.0.1:5000/sync
 ..
+
+4. Send REST request for running appy files to system:
+
+.. sourcecode:: console
+
+    $ curl -X POST http://127.0.0.1:5000/apply
+..
