+++ abc_server/flask/routes.py
@@ -1,11 +1,8 @@
-import threading
-
-from abc_server import auth
-from abc_server import config
 from abc_server.flask import abc_server
 from abc_server.flask import jsonify
-from abc_server.git import client as git_client
-from abc_server import utils
+from abc_server.service import apply_files
+from abc_server.service import base
+from abc_server.service import sync_files
 
 
 @abc_server.route('/')
@@ -15,7 +12,7 @@ def index():
 
 @abc_server.route('/init', methods=['POST'])
 def init():
-    git = get_git_client(init=True)
+    git = base.get_git_client(init=True)
     git.clone()
 
     return return_200()
@@ -23,28 +20,16 @@ def init():
 
 @abc_server.route('/sync', methods=['POST'])
 def sync():
-    sync_method()
+    sync_files.sync()
 
     return return_200()
 
 
-def get_git_client(init=False):
-    user = auth.authenticate()
-    if init:
-        repo = user.create_repo()
-    else:
-        repo = user.get_repo()
-    return git_client.GitClient(repo_url=repo.ssh_url)
+@abc_server.route('/apply', methods=['POST'])
+def apply():
+    apply_files.apply()
 
-
-def sync_method():
-    configs = config.read_config()
-    for item in configs['watch']:
-        utils.copy_files(item)
-    git = get_git_client()
-    git.commit()
-    git.push()
-    threading.Timer(24 * 3600, sync_method).start()
+    return return_200()
 
 
 def return_200(**kwargs):
