+++ abc_server/service/apply_files.py
@@ -1,0 +1,8 @@
+from abc_server import config
+from abc_server import utils
+
+
+def apply():
+    configs = config.read_config()
+    for item in configs['watch']:
+        utils.copy_files(item, sync=False)
