+++ abc_server/utils.py
@@ -15,10 +15,14 @@ def read_from_file(file_name):
         return yaml.load(f)
 
 
-def copy_files(file_path):
+def copy_files(file_path, sync=True):
     file_name = os.path.basename(file_path)
+    copy_from = file_path
     if '~' in file_path:
-        shutil.copy2(from_home_dir(file_path[2:]),
-                     from_home_dir('.abc/{}'.format(file_name)))
-        return
-    shutil.copy2(file_path, from_home_dir('.abc/{}'.format(file_name)))
+        copy_from = from_home_dir(file_path[2:])
+    copy_to = from_home_dir('.abc/{}'.format(file_name))
+
+    if not sync:
+        copy_to, copy_from = copy_from, copy_to
+
+    shutil.copy2(copy_from, copy_to)
